requirements.txt
ndspy==3.0.0
skytemple-files==1.1.5
skytemple-dtef==1.1.1
skytemple-icons==1.1.0
pygobject==3.36.0
pycairo==1.18.2; sys_platform != "win32"
natsort==7.1.1
tilequant==0.4.0
skytemple-ssb-debugger==1.1.2
skytemple-eventserver >= 1.0.0
pypresence==4.2.0
pygal==2.4.0
CairoSVG==2.5.1
